Created separate file for constants & functions
