+++ src/components/secrets/Secret.tsx
@@ -36,227 +36,7 @@ import { ReactComponent as Trash } from '../../assets/icons/ic-delete.svg'
 import { KeyValueFileInput } from '../util/KeyValueFileInput'
 import '../configMaps/ConfigMap.scss'
 import { decode } from '../../util/Util'
-import { ReactComponent as InfoIcon } from '../../assets/icons/ic-info-i-outline.svg'
-import { getCustomOptionSelectionStyle } from '../v2/common/ReactSelect.utils'
-
-export const sampleJSONs = {
-    ESO_GoogleSecretsManager: {
-        secretStore: {
-            gcpsm: {
-                auth: {
-                    secretRef: {
-                        secretAccessKeySecretRef: {
-                            name: 'gcpsm-secret',
-                            key: 'secret-access-credentials',
-                        },
-                    },
-                },
-                projectID: 'myProject',
-            },
-        },
-        esoData: [
-            {
-                secretKey: 'prod-mysql-password',
-                key: 'secrets/prod-mysql-secrets',
-            },
-            {
-                secretKey: 'prod-mysql-password',
-                key: 'secrets/prod-mysql-secrets',
-            },
-        ],
-    },
-    ESO_AWSSecretsManager: {
-        secretStore: {
-            aws: {
-                service: 'SecretsManager',
-                region: 'us-east-1',
-                auth: {
-                    secretRef: {
-                        accessKeyIDSecretRef: {
-                            name: 'awssm-secret',
-                            key: 'access-key',
-                        },
-                        secretAccessKeySecretRef: {
-                            name: 'awssm-secret',
-                            key: 'secret-access-key',
-                        },
-                    },
-                },
-            },
-        },
-        esoData: [
-            {
-                secretKey: 'prod-mysql-password',
-                key: 'secrets/prod-mysql-secrets',
-                property: 'prodPassword',
-            },
-            {
-                secretKey: 'prod-mysql-password',
-                key: 'secrets/prod-mysql-secrets',
-                property: 'prodPassword',
-            },
-        ],
-    },
-    ESO_AzureSecretsManager: {
-        secretStore: {
-            azurekv: {
-                tenantId: 'd3bc2180-xxxx-xxxx-xxxx-154105743342',
-                vaultUrl: 'https://my-keyvault-name.vault.azure.net',
-                authSecretRef: {
-                    clientId: {
-                        name: 'azure-secret-sp',
-                        key: 'ClientID',
-                    },
-                    clientSecret: {
-                        name: 'azure-secret-sp',
-                        key: 'ClientSecret',
-                    },
-                },
-            },
-        },
-        esoData: [
-            {
-                secretKey: 'prod-mysql-password',
-                key: 'secrets/prod-mysql-secrets',
-            },
-            {
-                secretKey: 'prod-mysql-password',
-                key: 'secrets/prod-mysql-secrets',
-            },
-        ],
-    },
-    ESO_HashiCorpVault: {
-        secretStore: {
-            vault: {
-                server: 'http://my.vault.server:8200',
-                path: 'secret',
-                version: 'v2',
-                auth: {
-                    tokenSecretRef: {
-                        name: 'vault-token',
-                        key: 'token',
-                    },
-                },
-            },
-        },
-        esoData: [
-            {
-                secretKey: 'prod-mysql-password',
-                key: 'secrets/prod-mysql-secrets',
-                property: 'prodPassword',
-            },
-            {
-                secretKey: 'prod-mysql-password',
-                key: 'secrets/prod-mysql-secrets',
-                property: 'prodPassword',
-            },
-        ],
-    },
-    default: [
-        {
-            key: 'service/credentials',
-            name: 'secret-key',
-            property: 'property-name',
-            isBinary: true,
-        },
-        {
-            key: 'service/credentials',
-            name: 'secret-key',
-            property: 'property-name',
-            isBinary: true,
-        },
-    ],
-}
-
-export const dataHeaders = {
-    ESO_GoogleSecretsManager: (
-        <div>
-            #Sample Data<br />
-            #secretKey: Name of the secret<br />
-            #key: GCP secret name<br />
-            #property: GCP Secrets Manager secret key<br />
-            #secretAccessKeySecretRef.name: The secret name which would be used for authentication<br />
-            #secretAccessKeySecretRef.key: Key name containing SA key<br />
-            #projectID: GCP Project ID where secret is created<br />
-        </div>
-    ),
-    ESO_AWSSecretsManager: (
-        <div>
-            #Sample Data <br />
-            #accessKeyIDSecretRef.name: Name of secret created that would be used for authentication <br />
-            #region: The region where Secret is created <br />
-            #secretKey: Name of the secret created <br />
-            #key: AWS Secrets Manager secret name <br />
-            #property: AWS Secrets Manager secret key <br />
-        </div>
-    ),
-    ESO_AzureSecretsManager: (
-        <div>
-            #Sample Data <br />
-            #tenantId: azure tenant ID <br />
-            #vaultUrl: URL of your vault instance <br />
-            #authSecretRef.name: Name of secret created that would be used for authentication <br />
-            #secretKey: Name of the secret <br />
-            #key: Azure Key vault secret name <br />
-        </div>
-    ),
-    ESO_HashiCorpVault: (
-        <div>
-            #Sample Data <br />
-            #vault.server: Server URL where vault is running <br />
-            #vault.path: Path where secret is stored <br />
-            #tokenSecretRef.name: The secret name which would be used for authentication <br />
-            #tokenSecretRef.key: Key name containing token <br />
-            #secretKey: Name of the secret <br />
-            #key: Vault secret name <br />
-            #property: Vault secret key <br />
-        </div>
-    ),
-    default: (
-        <div>
-            # Sample Data<br></br># key: Secret key in backend<br></br># name: Name for this key in the generated secret
-            <br></br># property: Property to extract if secret in backend is a JSON object(optional)<br></br># isBinary: Set
-            this to true if configuring an item for a binary file stored(optional)<br></br>
-        </div>
-    )
-}
-
-
-export const getTypeGroups = (typeValue?: string) => {
-    const noGroups: any[] = [
-            { value: '', label: 'Kubernetes Secret' },
-            { value: 'KubernetesSecret', label: 'Mount Existing Kubernetes Secret' },
-        ],
-        esoGroups: any[] = [
-            { value: 'ESO_GoogleSecretsManager', label: 'Google Secrets Manager' },
-            { value: 'ESO_AWSSecretsManager', label: 'AWS Secrets Manager' },
-            { value: 'ESO_AzureSecretsManager', label: 'Azure Secrets Manager' },
-            { value: 'ESO_HashiCorpVault', label: 'Hashi Corp Vault' },
-        ],
-        ksoGroups: any[] = [
-            { value: 'AWSSecretsManager', label: 'AWS Secrets Manager', deprecated: true },
-            { value: 'AWSSystemManager', label: 'AWS System Manager', deprecated: true },
-            { value: 'HashiCorpVault', label: 'Hashi Corp Vault', deprecated: true },
-        ]
-
-    const externalType = [...noGroups, ...esoGroups, ...ksoGroups].find((x) => x.value === typeValue)
-
-    if (typeValue) return externalType
-    return [
-        {
-            label: '',
-            options: noGroups,
-        },
-        {
-            label: 'External Secret Operator (ESO)',
-            options: esoGroups,
-        },
-        {
-            label: 'Kubernetes External Secret (KES)',
-            options: ksoGroups,
-        },
-    ]
-}
+import { dataHeaders, getTypeGroups, GroupHeading, groupStyle, sampleJSONs, SecretFormProps, SecretOptions } from './secret.constants'
 
 const Secret = ({ respondOnSuccess, ...props }) => {
     const [appChartRef, setAppChartRef] = useState<{ id: number; version: string; name: string }>()
@@ -455,34 +235,6 @@ export function ListComponent({ icon = '', title, subtitle = '', onClick, classN
             {collapsible && <img className="configuration-list__arrow pointer" src={arrowTriangle} />}
         </article>
     )
-}
-
-interface EsoData {
-    secretKey: string,
-    key: string,
-    property: string
-}
-interface SecretFormProps {
-    id: number
-    appChartRef: { id: number; version: string; name: string }
-    appId: number
-    roleARN: string
-    name: string
-    index: number
-    external: boolean
-    externalType: string
-    secretData: { key: string; name: string; property: string; isBinary: boolean }[]
-    esoSecretData?: {secretStore: any, esoData: EsoData[] }
-    type: string
-    data: { k: string; v: string }[]
-    isUpdate: boolean
-    mountPath: string
-    keyValueEditable?: boolean
-    filePermission: string
-    subPath: boolean
-    update: (...args) => void
-    collapse: (...args) => void
-    initialise?: () => void
 }
 
 export const SecretForm: React.FC<SecretFormProps> = function (props) {
@@ -977,48 +729,11 @@ export const SecretForm: React.FC<SecretFormProps> = function (props) {
                     options={getTypeGroups()}
                     defaultValue={externalType && externalType !== '' ? getTypeGroups(externalType) : getTypeGroups()[0].options[0]}
                     onChange={onChange}
-                    styles={{
-                        ...multiSelectStyles,
-                        menu: (base) => ({ ...base, zIndex: 9999, textAlign: 'left' }),
-                        control: (base) => ({...base, border: '1px solid #d6dbdf', width: '450px'}),
-                        group: (base) => ({
-                            ...base,
-                            paddingTop: 0,
-                            paddingBottom: 0,
-                        }),
-                        groupHeading: (base) => ({
-                            ...base,
-                            fontWeight: 600,
-                            fontSize: '12px',
-                            height: '28px',
-                            color: 'var(--N900)',
-                            backgroundColor: 'var(--N100)',
-                            marginBottom: 0,
-                        }),
-                        indicatorsContainer: (provided, state) => ({
-                            ...provided,
-                        })
-                    }}
+                    styles={groupStyle()}
                     components={{
                         IndicatorSeparator: null,
-                        Option: (props) => {
-                            props.selectProps.styles.option = getCustomOptionSelectionStyle()
-                            return <components.Option {...props}>
-                                <div>
-                                    <div>{props.data.label}</div>
-                                    {props.data?.deprecated && <div className='cy-7 fw-4 fs-11'>Deprecated</div>}
-                                </div>
-                            </components.Option>
-                        },
-                        GroupHeading: (props) => {
-                            if(!props.data.label)return null;
-                            return <components.GroupHeading {...props}>
-                                <div className='flex flex-justify h-100'>
-                                    {props.data.label}
-                                    <a className='flex' href='https://github.com/external-secrets/external-secrets' target='_blank'><InfoIcon className='icon-dim-20' /></a>
-                                </div>
-                            </components.GroupHeading>
-                        }
+                        Option: SecretOptions,
+                        GroupHeading
                     }}
                     />
                 </div>
